requirements.txt
fastapi
uvicorn
pymongo
python-multipart
pdfplumber
sentence-transformers
faiss-cpu
openai
python-dotenv
